+++ spec/client/realm_client_spec.rb
@@ -97,18 +97,49 @@ RSpec.describe KeycloakAdmin::RealmClient do
     end
   end
 
+  describe "#save" do
+    let(:realm_name) { "valid-realm" }
+    let(:realm) { KeycloakAdmin::RealmRepresentation.from_hash(
+      "id" => realm_name,
+      "realm" => realm_name
+    )}
+
+    before(:each) do
+      @realm_client = KeycloakAdmin.realm(nil)
+
+      stub_token_client
+
+      expect_any_instance_of(RestClient::Resource).to receive(:post).with(realm.to_json, anything)
+    end
+
+    it "saves a realm" do
+      @realm_client.save(realm)
+    end
+
+    it "passes rest client options" do
+      rest_client_options = {verify_ssl: OpenSSL::SSL::VERIFY_NONE}
+      allow_any_instance_of(KeycloakAdmin::Configuration).to receive(:rest_client_options).and_return rest_client_options
+
+      expect(RestClient::Resource).to receive(:new).with(
+        "http://auth.service.io/auth/admin/realms", rest_client_options).and_call_original
+
+      @realm_client.save(realm)
+    end
+  end
+
   describe "#update" do
     let(:realm_name) { "valid-realm" }
+    let(:realm_json) { { smtpServer: { host: 'test_host' } } }
 
     before(:each) do
       @realm_client = KeycloakAdmin.realm(realm_name)
 
       stub_token_client
+      expect_any_instance_of(RestClient::Resource).to receive(:put).with(realm_json.to_json, anything)
     end
 
-    it "updates realm" do
-      expect_any_instance_of(RestClient::Resource).to receive(:put).with('{"smtpServer":{"host":"test_host"}}', anything)
-      @realm_client.update({ smtpServer: { host: 'test_host' } })
+    it "updates a realm" do
+      @realm_client.update(realm_json)
     end
 
     it "passes rest client options" do
@@ -118,8 +149,7 @@ RSpec.describe KeycloakAdmin::RealmClient do
       expect(RestClient::Resource).to receive(:new).with(
         "http://auth.service.io/auth/admin/realms/valid-realm", rest_client_options).and_call_original
 
-      expect_any_instance_of(RestClient::Resource).to receive(:put).with('{"smtpServer":{"host":"test_host"}}', anything)
-      @realm_client.update({ smtpServer: { host: 'test_host' } })
+      @realm_client.update(realm_json)
     end
   end
 end