+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * List clients
 * List groups, create/save a group
 * List roles, save a role
-* List realms, update/delete a realm
+* List realms, save/update/delete a realm
 * Support passing rest client options for user save and search
 * Support using gem without ActiveSupport